+++ src/plone/app/mosaic/forms.py
@@ -12,19 +12,25 @@ class MosaicDefaultAddForm(add.DefaultAddForm):
         'IVersionable.changeNote'
     ]
 
-    def updateFieldsFromSchemata(self):
-        super(MosaicDefaultAddForm, self).updateFieldsFromSchemata()
+    @property
+    def mosaic_enabled(self):
         schemata = [s for s in self.additionalSchemata]
 
         if ILayoutAware not in schemata:
             # if it is not a mosaic add form, carry on...
-            return
+            return False
 
         # so we have it in the schemata, but is the view the default view
         # for the content type?
         portal_types = getToolByName(self.context, 'portal_types')
         ptype = portal_types[self.portal_type]
         if ptype.default_view != 'layout_view':
+            return False
+        return True
+
+    def updateFieldsFromSchemata(self):
+        super(MosaicDefaultAddForm, self).updateFieldsFromSchemata()
+        if not self.mosaic_enabled:
             return
 
         # we do not want the extra groups...
@@ -36,6 +42,8 @@ class MosaicDefaultAddForm(add.DefaultAddForm):
                 self.fields = self.fields.omit(field_name)
 
     def nextURL(self):
+        if not self.mosaic_enabled:
+            return super(MosaicDefaultAddForm, self).nextURL()
         # very hacky way to get url of object created.
         # only other way would be override a bunch of code I do not
         # feel very comfortable overridding